% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hrrr.r
\name{download_hrrr}
\alias{download_hrrr}
\title{download_hrrr
download HRRR files using the herbie python.
To enable, must run vpRm::setup_herbie()
herbie github: https://github.com/blaylockbk/Herbie}
\usage{
download_hrrr(times = NULL, vpRm = NULL, hrrr_dir = ".")
}
\arguments{
\item{times}{(POSIXct):}

\item{vpRm}{(vpRm, character): An object of class vpRm, or a file path to a vpRm control file.}

\item{hrrr_dir}{(chr): the directory in which to save the downloaded hrrr data files}
}
\description{
download_hrrr
download HRRR files using the herbie python.
To enable, must run vpRm::setup_herbie()
herbie github: https://github.com/blaylockbk/Herbie
}
