% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_templ.r
\name{gen_templ}
\alias{gen_templ}
\title{Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.}
\usage{
gen_templ(matchdomain = NULL, lc_filename, vpRm_dir = ".", verbose = F)
}
\arguments{
\item{matchdomain}{(chr) filepath(s) of geospatial data files readable by terra::rast to run vpRm over.}

\item{lc_filename}{(chr) landcover data filepath}

\item{vpRm_dir}{(chr) filepath of the vpRm directory}

\item{verbose}{(bool) Print messages?}
}
\description{
Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.
}
