% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{Get_Stilt_Out_Filenames}
\alias{Get_Stilt_Out_Filenames}
\title{Get_Stilt_Out_Filenames
Get the footprint or trajectory filenames from a STILT output
Taken directly from R package AtmosfeaR <3}
\usage{
Get_Stilt_Out_Filenames(stilt_out_dir, outtype)
}
\arguments{
\item{stilt_out_dir}{(chr): the path to a STILT out directory}

\item{outtype}{(chr): One of ("foot","traj"). "foot" --> get the footprint output.  "traj" --> get the traj output.}
}
\description{
Get_Stilt_Out_Filenames
Get the footprint or trajectory filenames from a STILT output
Taken directly from R package AtmosfeaR <3
}
