% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green.r
\name{green}
\alias{green}
\title{green
Create a green up/down pair of rasters for use in masking gee to the growing season}
\usage{
green(
  evi,
  up_func = function(evi) {     return(evi > (0.5 * max(evi))) },
  down_func = function(evi) {     return(evi < (0.65 * max(evi))) }
)
}
\arguments{
\item{evi}{(SpatRaster): A full year's worth of EVI data}

\item{up_func}{(function): a function that takes an evi time series over a year and returns a boolean: is this pixel on this day past the greenup day?}

\item{down_func}{(function): a function that takes an evi time series over a year and returns a boolean: is this pixel on this day before the greendown day?}
}
\description{
green
Create a green up/down pair of rasters for use in masking gee to the growing season
}
