% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_vpRm.r
\name{init_vpRm}
\alias{init_vpRm}
\title{vpRm_init
Initialize directory structure for vpRm}
\usage{
init_vpRm(directory)
}
\arguments{
\item{directory}{(chr) filepath of the root of the vpRm directory tree}
}
\description{
vpRm_init
Initialize directory structure for vpRm
}
