% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_vpRm.r
\name{new_vpRm}
\alias{new_vpRm}
\title{new_vpRm
Initialize an object of class vpRm}
\usage{
new_vpRm(
  vpRm_dir = ".",
  lc_dir = NULL,
  isa_dir = NULL,
  temp_dir = NULL,
  par_dir = NULL,
  evi_dir = NULL,
  verbose = F
)
}
\arguments{
\item{vpRm_dir}{(chr): path to initialize vpRm into; where to store processed drivers and outputs}

\item{lc_dir}{(chr): path to land cover data}

\item{isa_dir}{(chr): path to impermeable surface data}

\item{temp_dir}{(chr): path to temperature data}

\item{par_dir}{(chr): path to initialize photosynthetically available radiation data}

\item{evi_dir}{(chr): path to vegetation index data}

\item{verbose}{(bool): print intermediary updates?}
}
\description{
new_vpRm
Initialize an object of class vpRm
}
