% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{parse_herbie_hrrr_times}
\alias{parse_herbie_hrrr_times}
\title{parse_herbie_hrrr_times}
\usage{
parse_herbie_hrrr_times(hrrr_filenames)
}
\arguments{
\item{hrrr_filenames}{(chr): hrrr modis filenames to extract date times from}
}
\description{
parse_herbie_hrrr_times
}
