% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{parse_modis_evi_times}
\alias{parse_modis_evi_times}
\title{parse_modis_evi_times}
\usage{
parse_modis_evi_times(evi_filenames)
}
\arguments{
\item{evi_filenames}{(chr): evi modis filenames to extract date times from}
}
\description{
parse_modis_evi_times
}
