% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.r
\name{parse_stilt_times}
\alias{parse_stilt_times}
\title{parse_stilt_times}
\usage{
parse_stilt_times(stilt_filenames)
}
\arguments{
\item{stilt_filenames}{(chr): hrrr modis filenames to extract date times from}
}
\description{
parse_stilt_times
}
