% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_2d.r
\name{proc_2d}
\alias{proc_2d}
\title{proc_2d
Convert 2d data to match crs and extent to template}
\usage{
proc_2d(driver, plate)
}
\arguments{
\item{driver}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for  land cover data file to be processed}

\item{plate}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for vpRm template}
}
\description{
proc_2d
Convert 2d data to match crs and extent to template
}
