% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_drivers.r
\name{proc_drivers}
\alias{proc_drivers}
\title{proc_drivers()
Process the driver data for a VPRM model}
\usage{
proc_drivers(vpRm, n_cores = 1)
}
\arguments{
\item{vpRm}{(vpRm) a vpRm object}

\item{n_cores}{(int): number of cores for parallel processing. vpRm only parallellizes over hourly times, not over spatial subsets}
}
\description{
proc_drivers()
Process the driver data for a VPRM model
}
