% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vpRm.r
\name{proc_drivers.vpRm}
\alias{proc_drivers.vpRm}
\title{process.vpRm()
Process the driver data for a VPRM model
calls functions proc*}
\usage{
proc_drivers.vpRm(vpRm)
}
\arguments{
\item{vpRm}{(vpRm) a vpRm object}
}
\description{
process.vpRm()
Process the driver data for a VPRM model
calls functions proc*
}
