% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_interp.r
\name{proc_interp}
\alias{proc_interp}
\title{proc_interp
process driver data that must be interpolated to fit to the time points of the vpRm template}
\usage{
proc_interp(driver, plate)
}
\arguments{
\item{driver}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for  land cover data file to be processed}

\item{plate}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for vpRm template}
}
\description{
proc_interp
process driver data that must be interpolated to fit to the time points of the vpRm template
}
