% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_lc.r
\name{proc_lc}
\alias{proc_lc}
\title{proc_lc
Convert Land cover data to match crs and extent to template}
\usage{
proc_lc(driver_filename, templ_filename, proc_filename)
}
\arguments{
\item{driver_filename}{(chr) file path to land cover data file to be processed}
}
\description{
proc_lc
Convert Land cover data to match crs and extent to template
}
