% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_simple_3d.r
\name{proc_simple_3d}
\alias{proc_simple_3d}
\title{proc_simple_3d
Convert 3d data to match crs and extent to template}
\usage{
proc_simple_3d(driver, plate)
}
\arguments{
\item{driver}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for  land cover data file to be processed}

\item{plate}{(Spat Raster or terra::SpatRaster): SpatRaster or filepath that can be input to terra::Rast for vpRm template}
}
\description{
proc_simple_3d
Convert 3d data to match crs and extent to template
}
