% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_domain.r
\name{set_domain}
\alias{set_domain}
\title{set_domain()}
\usage{
set_domain(
  vpRm,
  domain = NULL,
  vpRm_crs = NULL,
  vpRm_ext = NULL,
  vpRm_res = NULL,
  vpRm_time = NULL
)
}
\arguments{
\item{vpRm}{(vpRm): vpRm object you want to set the domain of}

\item{domain}{(spatRaster or chr): domain you want to run VRPM over}

\item{vpRm_crs}{(chr): Optional CRS.  Overwrites CRS of domain}

\item{vpRm_ext}{(vector): c(xmin, xmax, ymin, ymax)}

\item{vpRm_res}{(chr): Optional resolution. Overwrites res of domain}

\item{vpRm_time}{(vector): Optional times to run VPRM over. Overwrites time of domain

If not NULL- vpRm_ext, vpRm_crs, vpRm_time, will overwrite the CRS of domain, if domain is given.}
}
\description{
Sets the domain of a vpRm object
Give a vpRm object a temporal spatial domain over which to run VPRM.
Sets the output CRS and extent.
Can take a given terra::spatRaster object, or directory to geospatial data readable by terra.
For example, a STILT domain you want to convolve with the VPRM NEE output field
Can also take a projection, xy bounds, and and time series.
}
