% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_hrrr.r
\name{setup_herbie}
\alias{setup_herbie}
\title{setup_herbie
Set up herbie for downloading HRRR meteorology files.
An installation of conda is a prerequisite.
if this fails to automatically setup the
This is taken from the herbie github: https://github.com/blaylockbk/Herbie}
\usage{
setup_herbie()
}
\description{
setup_herbie
Set up herbie for downloading HRRR meteorology files.
An installation of conda is a prerequisite.
if this fails to automatically setup the
This is taken from the herbie github: https://github.com/blaylockbk/Herbie
}
