% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpRm.r
\name{vpRm}
\alias{vpRm}
\title{vpRm
Wrapper function to run a vpRm model from start to finish.}
\usage{
vpRm(
  vpRm_dir = ".",
  lc_dir = NULL,
  isa_dir = NULL,
  temp_dir = NULL,
  dswrf_dir = NULL,
  landsat_dir = NULL,
  evi_dir = NULL,
  lswi_dir = NULL,
  evi_extrema_dir = NULL,
  green_dir = NULL,
  vprm_params = vpRm::vprm_params,
  domain_template = NULL,
  verbose = F
)
}
\arguments{
\item{vpRm_dir}{(chr): path to initialize vpRm into; where to store processed drivers and outputs}

\item{lc_dir}{(chr): path to land cover data}

\item{isa_dir}{(chr): path to impermeable surface data}

\item{temp_dir}{(chr): path to temperature data}

\item{dswrf_dir}{(chr): path to initialize photosynthetically available radiation data}

\item{landsat_dir}{(chr): path to raw landsat reflectance data}

\item{evi_dir}{(chr): path to vegetation index data}

\item{lswi_dir}{(chr): path to LSWI data}

\item{evi_extrema_dir}{(chr): path to evi_extrema 2d data}

\item{green_dir}{(chr): path to greenup/down data}

\item{vprm_params}{(data.frame): VPRM parameters, trained from tower data}

\item{domain_template}{(chr, terra::spatRaster): Spatiotemporal domain over which to run VPRM. For example: a series of STILT footprints.}

\item{verbose}{(bool): print intermediary updates?}
}
\description{
vpRm
Wrapper function to run a vpRm model from start to finish.
}
