% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm_params.r
\docType{data}
\name{vprm_params}
\alias{vprm_params}
\title{VPRM parameters}
\format{
Data frame with 11 rows and 9 cols.
\describe{
lc
item{ Tmin }
item{ Topt }
item{ Tmax }
item{ Tlow }
item{ PAR0 }
item{ lambda }
item{ alpha }
item{ beta }
}
}
\usage{
vprm_params
}
\description{
VPRM parameters for different landcover types. LC codes match NA landcover database
}
\keyword{datasets}
