% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vprm_params_nlcd.r
\docType{data}
\name{vprm_params_nlcd}
\alias{vprm_params_nlcd}
\title{VPRM parameters (NLCD)}
\format{
Data frame with 11 rows and 9 cols.
\describe{
lc
item{ Tmin }
item{ Topt }
item{ Tmax }
item{ Tlow }
item{ PAR0 }
item{ lambda }
item{ alpha }
item{ beta }
}
}
\usage{
vprm_params_nlcd
}
\description{
VPRM parameters for different landcover types. LC codes match NLCD landcover database.
}
\keyword{datasets}
