% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_templ.r
\name{gen_plate}
\alias{gen_plate}
\title{gen_plate()
Creates an empty terra rast template to match other processed driver data to.
Extent of the input match domain with the projection of the landcover driver data}
\usage{
gen_plate(matchdomain = NULL, lc_filename, vpRm_dir = ".", verbose = F)
}
\arguments{
\item{matchdomain}{(chr) filepath(s) of geospatial data files readable by terra::rast to run vpRm over.}

\item{lc_filename}{(chr) landcover data filepath}

\item{vpRm_dir}{(chr) filepath of the vpRm directory}

\item{verbose}{(bool) Print messages?}
}
\description{
gen_plate()
Creates an empty terra rast template to match other processed driver data to.
Extent of the input match domain with the projection of the landcover driver data
}
