% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_templ.r
\name{gen_templ}
\alias{gen_templ}
\title{Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.}
\usage{
gen_templ(
  matchdomain = NULL,
  spatextect,
  xres = NULL,
  yres = NULL,
  times = NULL,
  lc_filename,
  vpRm_dir = ".",
  outfile,
  verbose = F
)
}
\arguments{
\item{matchdomain}{}

\item{times}{\subsection{NOT IMPLEMENTED}{
}}

\item{outfile}{\subsection{NOT IMPLEMENTED}{
}}

\item{lons}{\subsection{NOT IMPLEMENTED}{
}}

\item{lats}{\subsection{NOT IMPLEMENTED}{
}}

\item{projection}{\subsection{NOT IMPLEMENTED}{
}}
}
\description{
Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.
}
