% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_raster_templ.r
\name{gen_templ}
\alias{gen_templ}
\title{Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.}
\usage{
gen_templ(domain, lons, lats, times)
}
\arguments{
\item{domain}{}

\item{lons}{}

\item{lats}{}

\item{times}{}
}
\description{
Gen_Raster_Templ()
Creates the raster brick/stack/ netCDF4 structure that will be filled with processed driver data and VPRM NEE output.
}
