% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_vpRm.r
\name{new_vpRm}
\alias{new_vpRm}
\title{new_vpRm
Initialize an object of class vpRm}
\usage{
new_vpRm(
  vpRm_dir = ".",
  lc_dir = NULL,
  isa_dir = NULL,
  temp_dir = NULL,
  dswrf_dir = NULL,
  landsat_dir = NULL,
  evi_dir = NULL,
  lswi_dir = NULL,
  evi_extrema_dir = NULL,
  green_dir = NULL,
  vprm_params = vpRm::vprm_params,
  vpRm_crs = NULL,
  vpRm_ext = NULL,
  vpRm_res = NULL,
  vpRm_time = NULL,
  verbose = F
)
}
\arguments{
\item{vpRm_dir}{(chr): path to initialize vpRm into; where to store processed drivers and outputs}

\item{lc_dir}{(chr): path to land cover data}

\item{isa_dir}{(chr): path to impermeable surface data}

\item{temp_dir}{(chr): path to temperature data}

\item{dswrf_dir}{(chr): path to initialize photosynthetically available radiation data}

\item{landsat_dir}{(chr): path to raw landsat reflectance data}

\item{evi_dir}{(chr): path to vegetation index data}

\item{lswi_dir}{(chr): path to LSWI data}

\item{evi_extrema_dir}{(chr): path to evi_extrema 2d data}

\item{green_dir}{(chr): path to greenup/down data}

\item{vprm_params}{(data.frame): VPRM parameters, trained from tower data}

\item{vpRm_crs}{(terra::crs): coordinate reference system for intake}

\item{vpRm_ext}{(terra::ext): extent for intake}

\item{vpRm_res}{(terra::ext): resolution for intake}

\item{vpRm_time}{(terra::time): times of intake}

\item{verbose}{(bool): print intermediary updates?}
}
\description{
new_vpRm
Initialize an object of class vpRm
}
