% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_drivers.r
\name{proc_drivers}
\alias{proc_drivers}
\title{proc_drivers.vpRm()
Process the driver data for a VPRM model
calls functions proc*}
\usage{
proc_drivers(vpRm)
}
\arguments{
\item{vpRm}{(vpRm) a vpRm object}
}
\description{
proc_drivers.vpRm()
Process the driver data for a VPRM model
calls functions proc*
}
