% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.r
\name{run_vpRm}
\alias{run_vpRm}
\title{run_vpRm
Run a VPRM model defined by a vpRm object}
\usage{
run_vpRm(vpRm)
}
\arguments{
\item{vpRm}{(vpRm): a vpRm S3 object with attached driver data}
}
\value{
vpRm (vpRm): the same vpRm object, now with attached gee, respiration and nee netcdf files.
}
\description{
Execute the models calculations defined in Mahadevan et al 2008 and Winbourne et al 2021.  Processes the driver data attached to the vpRm object with a call to proc_drivers().
}
