% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{Save_Rast}
\alias{Save_Rast}
\title{Save_Rast
Save a SpatRaster as a netcdf}
\usage{
Save_Rast(SpatRaster, filename)
}
\arguments{
\item{SpatRaster}{(spatRaster): SpatRaster to be saved}

\item{filename}{(chr): filename to save to}
}
\description{
Save_Rast
Save a SpatRaster as a netcdf
}
