% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{Save_Rast}
\alias{Save_Rast}
\title{Save_Rast
Save a SpatRaster as a netcdf}
\usage{
Save_Rast(RAST, filename)
}
\arguments{
\item{filename}{(chr): filename to save to}

\item{rast}{(SpatRaster): SpatRaster to be saved}
}
\description{
Save_Rast
Save a SpatRaster as a netcdf
}
